requirements.txt
opencv-python
matplotlib
Pillow
numpy